+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `FastIca`, an independent component analysis (ICA) algorithm.
+
 ## [0.2.0] - 2020-05-24
 
 ### Added
@@ -28,6 +34,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Principal component analysis (PCA).
 
+[Unreleased]: https://github.com/petabi/petal-decomposition/compare/0.2.0...master
 [0.2.0]: https://github.com/petabi/petal-decomposition/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/petabi/petal-decomposition/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/petabi/petal-decomposition/tree/0.1.0